python-backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-jobspy==1.11.8
pandas==2.1.4
gunicorn==21.2.0